added dashboard application and transfared all the views and urls from frontPage to dashboard
