+++ Rizu/urls.py
@@ -20,5 +20,5 @@ from django.urls import include, path
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("frontPage/", include("frontPage.urls")),
+    path("dashboard/", include("dashboard.urls")),
 ]
